requirements.txt
instaloader
python-dotenv
requests